Very close to full working v1

+++ src/procs/monitor.py
@@ -1,50 +1,53 @@
 import time, os, sys, signal, subprocess
-from .spotify import current_album_name, current_track_name, current_track_id, pause, get
-from .file    import get_path
-from .data 	  import load_from_local, save_to_local
+from ipc.terminal import send_signal, wake_on_signal_received
+from lib.spotify  import current, get, get_spotify_permissions
+from filesys.file   import get_path, music_home
+from lib.data 	  import load_from_local, save_to_local
+from ipc.terminal import prompt_and_get_rating
 
-def create_monitor(tid):
-	# to_zap = os.fork()
-	# if to_zap == 0:
-	# 	# Child to get zapped
-	# 	try:
-	# 		time.sleep(100000)
-	# 	except:
-	# 		sys.exit()
-	if os.fork() == 0:
-		# if os.fork() > 0:
-		# 	sys.exit()
-		set_monitor_pid(os.getpid())
-		# os.kill(to_zap, signal.SIGINT)
-		monitor(tid)
-	# os.waitpid(to_zap,0)
-
-def monitor(tid):
+def monitor():
 	prev_track = None
 	prev_album = None
+	signal.signal(signal.SIGUSR1, signal.SIG_IGN)
 	while True:
-		cur_track = current_track_name()
-		cur_album = current_album_name()
-		if cur_track != prev_track:
+		cur_track = current()
+		cur_album = None
+		if cur_track != None:
 			if prev_track != None:
-				info = open('/Users/spencersharp/.spools/music/info.txt','w')
-				info.write('rate -dp -s ' + current_track_id() )#+ ' | /dev/tty')
-				info.close()
-				print('\nPlease score {}: '.format(cur_track), end='', flush=True)
-				pause()
-				#os.kill(tid, signal.SIGUSR1)
-				# os.open(tid)
-				# #os.system(cmd)
-				# os.close(tid)
-				#subprocess.call(["python"],path,"rate","-dip","-s", cur_track],shell=True)
-			prev_track = cur_track
-		if cur_album != prev_album:
+				if cur_track.id != prev_track.id:
+					prompt_and_get_rating(prev_track)
+					prev_track = cur_track
+			else:
+				prev_track = cur_track
+		if False and cur_album != None:
+			cur_album = get('album', cur_track.album_id)
 			if prev_album != None:
-				cmd =  'python /Users/spencersharp/Documents/Coding/Active/spools-music/src/music.py'
-				cmd += ' rate -dp -a ' + cur_album
-				#subprocess.call(cmd)
-			prev_album = cur_album
+				if cur_album.id != prev_album.id:
+					prompt_and_get_rating(prev_album)
+					prev_album = cur_album
+			else:
+				prev_album = cur_album
 		time.sleep(0.1)
+
+def create_monitor():
+	top_fork = os.fork()
+	if top_fork == 0:
+		signal.pause()
+		print('Top exiting',flush=True)
+		sys.exit()
+	signal.signal(signal.SIGUSR1, wake_on_signal_received)
+	blocked = open(music_home + 'monitor_block', 'w')
+	if os.fork() == 0:
+		blocked.close()
+		blocked = open(music_home + 'monitor_block', 'w')
+		if os.fork() > 0:
+			sys.exit()
+		
+		set_monitor_pid(os.getpid())
+		send_signal(top_fork,signal.SIGUSR1)
+		monitor()
+	return top_fork, blocked
+	
 
 def get_monitor_path():
 	return get_path('monitor')
@@ -66,6 +69,8 @@ def get_monitor_pid():
 
 def does_monitor_exist():
 	pid = get_monitor_pid()
+	if pid == -1:
+		return False
 	try:
 		os.kill(pid, 0)
 	except:
